+++ components/script/dom/domparser.rs
@@ -51,7 +51,7 @@ impl<'a> DOMParserMethods for JSRef<'a, DOMParser> {
                        -> Fallible<Temporary<Document>> {
         let window = self.window.root();
         let url = window.r().get_url();
-        let content_type = DOMParserBinding::SupportedTypeValues::strings[ty as uint].to_owned();
+        let content_type = DOMParserBinding::SupportedTypeValues::strings[ty as usize].to_owned();
         match ty {
             Text_html => {
                 let document = Document::new(window.r(), Some(url.clone()),
