+++ components/script/dom/worker.rs
@@ -33,7 +33,7 @@ pub type TrustedWorkerAddress = Trusted<Worker>;
 #[dom_struct]
 pub struct Worker {
     eventtarget: EventTarget,
-    refcount: Cell<uint>,
+    refcount: Cell<usize>,
     global: GlobalField,
     /// Sender to the Receiver associated with the DedicatedWorkerGlobalScope
     /// this Worker created.
